+++ modules/signal_processor.py
